created widget folder and separate card components
