+++ app/lib/app/(public)/request_details/widgets/request_details_card/request_details_card_created_at_date.dart
@@ -1,0 +1,33 @@
+import 'package:core_module/core_module.dart';
+import 'package:flutter/material.dart';
+
+class RequestDetailsCardCreatedAtDate extends StatelessWidget {
+  const RequestDetailsCardCreatedAtDate({
+    super.key,
+    required this.createdAt,
+  });
+
+  final DateTime createdAt;
+
+  @override
+  Widget build(BuildContext context) {
+    final textThemes = Theme.of(context).textTheme;
+    return Row(
+      children: [
+        Icon(
+          UIcons.regularRounded.clock,
+          size: 16.0,
+          color: const Color(0xff767373), //TODO: atualizar com a cor correta
+        ),
+        const SizedBox(width: 10.0),
+        Text(
+          DateAdapter.dateToString(createdAt),
+          style: textThemes.bodySmall?.copyWith(
+            fontSize: 14.0,
+            color: const Color(0xff767373), //TODO: atualizar com a cor correta
+          ),
+        ),
+      ],
+    );
+  }
+}
